% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct.R
\name{ct}
\alias{ct}
\title{A Count Function}
\usage{
ct(df, var)
}
\arguments{
\item{df}{Name of the Dataset}

\item{var}{Variable to Count}
}
\description{
This function gives you a simple count with percentages
}
\examples{
ct()
}
\keyword{Count}
