% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_wt.R
\name{ct_wt}
\alias{ct_wt}
\title{A Count Function with Weights}
\usage{
ct_wt(df, var, wt)
}
\arguments{
\item{df}{Name of the Dataset}

\item{var}{Variable to Count}

\item{wt}{Weighting Variable}
}
\description{
This function gives you a simple count with percentages that is weighted
}
\examples{
ct_wt()
}
\keyword{Count}
