% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_ci.R
\name{mean_ci}
\alias{mean_ci}
\title{A Mean with 95 percent CIs}
\usage{
mean_ci(df, var)
}
\arguments{
\item{df}{Name of the Dataset}

\item{var}{Variable to find the mean of}
}
\description{
This function gives you a mean with 95 percent CIs
}
\examples{
mean_ci()
}
\keyword{Mean}
