% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_pad.R
\name{mean_pad}
\alias{mean_pad}
\title{A Mean with 84 percent CIs}
\usage{
mean_pad(df, var)
}
\arguments{
\item{df}{Name of the Dataset}

\item{var}{Variable to find the mean of}
}
\description{
This function gives you a mean with 84 percent CIs
}
\examples{
mean_pad()
}
\keyword{Mean}
