% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_col.R
\name{gg_col}
\alias{gg_col}
\title{A Quick ggcol chart}
\usage{
gg_col(df, xvar, yvar)
}
\arguments{
\item{df}{Name of the Dataset}

\item{xvar}{What is going to be on the x-axis}

\item{yvar}{What is going to be on the y-axis, usually pct}
}
\description{
This function gives you a histogram with percentage annotations
}
\examples{
gg_col()
}
\keyword{plotting}
